demo_pytest_02/requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
atomicwrites==1.4.0
attrs==21.4.0
colorama==0.4.4
importlib-metadata==4.11.3
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.8
pytest==7.1.1
pytest-html==3.1.1
pytest-metadata==2.0.1
six==1.16.0
tomli==2.0.1
typing-extensions==4.1.1
zipp==3.8.0
